+++ packages/editor/editor-core/__tests__/create-editor/create-editor.ts
@@ -78,7 +78,6 @@ describe(name, () => {
             () => {},
             {} as any,
             {} as any,
-            {} as any,
           ).length,
         ).toEqual(1);
       });