+++ packages/editor/editor-core/examples/2-comment.tsx
@@ -3,6 +3,7 @@
 import * as React from 'react';
 
 import Button, { ButtonGroup } from '@atlaskit/button';
+import LockCircleIcon from '@atlaskit/icon/glyph/lock-circle';
 import Editor from './../src/editor';
 import EditorContext from './../src/ui/EditorContext';
 import WithEditorActions from './../src/ui/WithEditorActions';
@@ -139,6 +140,13 @@ export default class EditorWithFeedback extends React.Component<Props, State> {
                     allowExtension={true}
                     insertMenuItems={customInsertMenuItems}
                     extensionHandlers={extensionHandlers}
+                    secondaryToolbarComponents={[
+                      <LockCircleIcon
+                        key="permission"
+                        size="large"
+                        label="Permissions"
+                      />,
+                    ]}
                   />
                 </CollapsedEditor>
               </div>