+++ packages/editor/renderer/src/email/nodes/bullet-list.ts
@@ -1,0 +1,10 @@
+import { createTag, serializeStyle } from '../util';
+import { NodeSerializerOpts } from '../interfaces';
+
+const css = serializeStyle({
+  'list-style-type': 'disc',
+});
+
+export default function bulletList({ text }: NodeSerializerOpts) {
+  return createTag('ul', { style: css }, text);
+}