+++ packages/editor/renderer/src/index.ts
@@ -10,6 +10,8 @@ import { Serializer } from './serializer';
 
 export { default as ReactSerializer } from './react';
 export { default as TextSerializer } from './text';
+export { default as EmailSerializer } from './email';
+
 export { default as ReactRenderer } from './ui/Renderer';
 export { RendererContext } from './react';
 export { ADFEncoder } from './utils';