+++ packages/editor/renderer/src/ui/index.tsx
@@ -1,3 +1,1 @@
 export { default as Renderer } from './Renderer';
-export { default as MediaItem } from './MediaItem';
-export { Appearance } from './MediaItem/MediaComponent';