+++ packages/media/media-card/src/app/styled/AppCardView.ts
@@ -17,6 +17,7 @@ import { borderRadius, cardShadow } from '../../styles';
 const previewWidth = 116;
 
 export interface CardProps {
+  isClickable: boolean;
   background: string | undefined;
 }
 
@@ -70,6 +71,16 @@ const cardOverlay = ({ background }: CardProps) => {
   }
 };
 
+const cardCursor = ({ isClickable }: CardProps) => {
+  if (isClickable) {
+    return `
+      cursor: pointer;
+    `;
+  } else {
+    return '';
+  }
+};
+
 export const Card: ComponentClass<HTMLAttributes<{}> & CardProps> = styled.div`
   ${cardColors} ${cardOverlay} display: inline-flex; /* make the card fit to its contents */
   flex-direction: row; /* make the preview and content side-by-side */
@@ -81,6 +92,7 @@ export const Card: ComponentClass<HTMLAttributes<{}> & CardProps> = styled.div`
   font-family: ${akFontFamily};
 
   ${borderRadius} ${cardShadow};
+  ${cardCursor};
 `;
 
 export interface PreviewProps {