+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-mobile-bridge
 
+## 0.0.5
+
+## 0.0.4
+
 ## 0.0.3
 - [patch] Fix dependency [d4bcbf4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d4bcbf4)